+++ src/storage/mod.rs
@@ -123,10 +123,7 @@ use crate::{
         metrics::{CommandKind, *},
         mvcc::{MvccReader, PointGetterBuilder},
         txn::{
-            commands::{
-                acquire_pessimistic_lock::PessimisticLockCmdInner, RawAtomicStore,
-                RawCompareAndSwap, TypedCommand,
-            },
+            commands::{RawAtomicStore, RawCompareAndSwap, TypedCommand},
             flow_controller::{EngineFlowController, FlowController},
             scheduler::Scheduler as TxnScheduler,
             Command, ErrorInner as TxnError,
@@ -1419,7 +1416,7 @@ impl<E: Engine, L: LockManager, F: KvFormat> Storage<E, L, F> {
         callback: Callback<T>,
     ) -> Result<()> {
         use crate::storage::txn::commands::{
-            AcquirePessimisticLock, Prewrite, PrewritePessimistic,
+            AcquirePessimisticLock, AcquirePessimisticLockResumed, Prewrite, PrewritePessimistic,
         };
 
         let cmd: Command = cmd.into();
@@ -1445,28 +1442,28 @@ impl<E: Engine, L: LockManager, F: KvFormat> Storage<E, L, F> {
                 )?;
                 check_key_size!(keys, self.max_key_size, callback);
             }
-            Command::AcquirePessimisticLock(AcquirePessimisticLock { inner, .. }) => match inner {
-                PessimisticLockCmdInner::SingleRequest { keys, .. } => {
-                    let keys = keys.iter().map(|k| k.0.as_encoded());
-                    Self::check_api_version(
-                        self.api_version,
-                        cmd.ctx().api_version,
-                        CommandKind::acquire_pessimistic_lock,
-                        keys.clone(),
-                    )?;
-                    check_key_size!(keys, self.max_key_size, callback);
-                }
-                PessimisticLockCmdInner::BatchResumedRequests { items, .. } => {
-                    let keys = items.iter().map(|item| item.key.as_encoded());
-                    Self::check_api_version(
-                        self.api_version,
-                        cmd.ctx().api_version,
-                        CommandKind::acquire_pessimistic_lock_resumed,
-                        keys.clone(),
-                    )?;
-                    check_key_size!(keys, self.max_key_size, callback);
-                }
-            },
+            Command::AcquirePessimisticLock(AcquirePessimisticLock { keys, .. }) => {
+                let keys = keys.iter().map(|k| k.0.as_encoded());
+                Self::check_api_version(
+                    self.api_version,
+                    cmd.ctx().api_version,
+                    CommandKind::prewrite,
+                    keys.clone(),
+                )?;
+                check_key_size!(keys, self.max_key_size, callback);
+            }
+            Command::AcquirePessimisticLockResumed(AcquirePessimisticLockResumed {
+                items, ..
+            }) => {
+                let keys = items.iter().map(|item| item.key.as_encoded());
+                Self::check_api_version(
+                    self.api_version,
+                    cmd.ctx().api_version,
+                    CommandKind::acquire_pessimistic_lock_resumed,
+                    keys.clone(),
+                )?;
+                check_key_size!(keys, self.max_key_size, callback);
+            }
             _ => {}
         }
         with_tls_tracker(|tracker| {
@@ -3346,7 +3343,7 @@ pub mod test_util {
     ) -> PessimisticLockCommand {
         let primary = keys[0].0.clone().to_raw().unwrap();
         let for_update_ts: TimeStamp = for_update_ts.into();
-        commands::AcquirePessimisticLock::new_disallow_lock_with_conflict(
+        commands::AcquirePessimisticLock::new(
             keys,
             primary,
             start_ts.into(),
@@ -3358,6 +3355,7 @@ pub mod test_util {
             for_update_ts.next(),
             OldValues::default(),
             check_existence,
+            false,
             false,
             Context::default(),
         )
@@ -8116,7 +8114,7 @@ mod tests {
         // Meet lock-k.
         storage
             .sched_txn_command(
-                commands::AcquirePessimisticLock::new_disallow_lock_with_conflict(
+                commands::AcquirePessimisticLock::new(
                     vec![(Key::from_raw(b"foo"), false), (Key::from_raw(&k), false)],
                     k.clone(),
                     20.into(),
@@ -8127,6 +8125,7 @@ mod tests {
                     false,
                     21.into(),
                     OldValues::default(),
+                    false,
                     false,
                     false,
                     Context::default(),
@@ -8208,7 +8207,7 @@ mod tests {
             for k in keys {
                 storage
                     .sched_txn_command(
-                        commands::AcquirePessimisticLock::new_disallow_lock_with_conflict(
+                        commands::AcquirePessimisticLock::new(
                             vec![(k.clone(), false)],
                             k.to_raw().unwrap(),
                             lock_ts.into(),
@@ -8219,6 +8218,7 @@ mod tests {
                             false,
                             (lock_ts + 1).into(),
                             OldValues::default(),
+                            false,
                             false,
                             false,
                             Context::default(),
@@ -8793,7 +8793,7 @@ mod tests {
         // T1 acquires lock on k1, start_ts = 1, for_update_ts = 3
         storage
             .sched_txn_command(
-                commands::AcquirePessimisticLock::new_disallow_lock_with_conflict(
+                commands::AcquirePessimisticLock::new(
                     vec![(key1.clone(), false)],
                     k1.to_vec(),
                     1.into(),
@@ -8806,6 +8806,7 @@ mod tests {
                     OldValues::default(),
                     false,
                     false,
+                    false,
                     Default::default(),
                 ),
                 expect_ok_callback(tx.clone(), 0),
@@ -8816,7 +8817,7 @@ mod tests {
         // T2 acquires lock on k2, start_ts = 10, for_update_ts = 15
         storage
             .sched_txn_command(
-                commands::AcquirePessimisticLock::new_disallow_lock_with_conflict(
+                commands::AcquirePessimisticLock::new(
                     vec![(key2.clone(), false)],
                     k2.to_vec(),
                     10.into(),
@@ -8827,6 +8828,7 @@ mod tests {
                     false,
                     0.into(),
                     OldValues::default(),
+                    false,
                     false,
                     false,
                     Default::default(),
@@ -9046,7 +9048,7 @@ mod tests {
                 ExpectedWrite::new().expect_proposed_cb(),
             ],
 
-            command: AcquirePessimisticLock::new_disallow_lock_with_conflict(
+            command: AcquirePessimisticLock::new(
                 keys.iter().map(|&it| (Key::from_raw(it), true)).collect(),
                 keys[0].to_vec(),
                 TimeStamp::new(10),
@@ -9057,6 +9059,7 @@ mod tests {
                 false,
                 TimeStamp::new(12),
                 OldValues::default(),
+                false,
                 false,
                 false,
                 Context::default(),
@@ -9072,7 +9075,7 @@ mod tests {
                 ExpectedWrite::new().expect_no_proposed_cb(),
             ],
 
-            command: AcquirePessimisticLock::new_disallow_lock_with_conflict(
+            command: AcquirePessimisticLock::new(
                 keys.iter().map(|&it| (Key::from_raw(it), true)).collect(),
                 keys[0].to_vec(),
                 TimeStamp::new(10),
@@ -9083,6 +9086,7 @@ mod tests {
                 false,
                 TimeStamp::new(12),
                 OldValues::default(),
+                false,
                 false,
                 false,
                 Context::default(),
